[feat]: Responsividade do projeto e criacao do README.md
