+++ my-app/src/styles/ContentWeather.css
@@ -1,9 +1,6 @@
 .container-content-weather {
   display: flex;
   flex-direction: column;
-  width: 100%;
-  height: 100%;
-  padding: 30px;
   background-color: #f5f5f5;
 }
 
@@ -12,17 +9,10 @@
   grid-template-columns: 2fr 2fr;
   grid-template-rows: 2fr 2fr;
   justify-items: start;
-  width: 70%;
-  height: 100%;
-  padding-top: 10px;
-  padding-bottom: 15px;
 }
 
 .__info__humidade {
   display: flex;
-  flex-direction: row;
-  justify-content: center;
-  align-items: center;
   width: 100%;
   height: 100%;
   gap: 1rem;
@@ -30,9 +20,6 @@
 
 .__info__sensacao {
   display: flex;
-  flex-direction: row;
-  justify-content: center;
-  align-items: center;
   width: 100%;
   height: 100%;
   gap: 1rem;
@@ -40,9 +27,6 @@
 
 .__info__vento {
   display: flex;
-  flex-direction: row;
-  justify-content: center;
-  align-items: center;
   width: 100%;
   height: 100%;
   gap: 1rem;
@@ -50,16 +34,13 @@
 
 .__info_valuesTemp {
   display: flex;
-  flex-direction: row;
-  justify-content: center;
-  align-items: center;
   width: 100%;
   height: 100%;
   gap: 1rem;
 }
 
 .local h2 {
-  font-size: 1rem;
+  font-size: 1.2rem;
   font-weight: 600;
   color: rgb(74, 73, 73);
 }
@@ -72,26 +53,34 @@
 
 .container__week {
   border-top: 2px solid #f4f40e;
-  width: 90%;
+  width: 100%;
   height: 100%;
   display: grid;
-  grid-template-columns: 1fr 1fr 1fr 1fr 1fr 1fr;
-  justify-items: center;
+  grid-template-columns: 1fr 1fr 1fr 1fr 1fr;
   padding: 10px;
 }
 
-.container__week__day__info {
-  display: flex;
-  flex-direction: row;
-  width: 100%;
-  height: 100%;
-  gap: 1rem;
+.local__container {
+  padding: 10px;
 }
 
-.container__week__day {
-  display: flex;
-  flex-direction: column;
-  width: 100%;
-  height: 100%;
-  gap: 1rem;
+.info_data {
+  font-size: 1rem;
+  font-weight: 700;
+  color: rgb(90, 90, 90);
+}
+
+.__info_title {
+  font-size: 1rem;
+  font-weight: 500;
+  color: rgb(90, 90, 90);
+}
+
+@media screen and (min-width: 600px) {
+  .container-content-weather {
+    display: flex;
+    flex-direction: column;
+    background-color: #f5f5f5;
+    padding: 30px;
+  }
 }
